move validation test to olm suite, update e2e tests
