+++ internal/scorecard/plugins/plugin_runner.go
@@ -290,7 +290,6 @@ func RunInternalPlugin(pluginType PluginType, config BasicAndOLMPluginConfig, lo
 			conf := BasicTestConfig{
 				Client:   runtimeClient,
 				CR:       obj,
-				Bundle:   config.Bundle,
 				ProxyPod: proxyPodGlobal,
 			}
 			basicTests := NewBasicTestSuite(conf)
@@ -314,6 +313,7 @@ func RunInternalPlugin(pluginType PluginType, config BasicAndOLMPluginConfig, lo
 				CSV:      csv,
 				CRDsDir:  config.CRDsDir,
 				ProxyPod: proxyPodGlobal,
+				Bundle:   config.Bundle,
 			}
 			olmTests := NewOLMTestSuite(conf)
 			if schelpers.IsV1alpha2(config.Version) {
